feat: move metrics middleware to middleware module
